refactor: rename gh_fetcher.rs to github_fetcher.rs
