+++ src/bin/clydetools/main.rs
@@ -11,7 +11,7 @@ use clap::{Parser, Subcommand};
 pub mod add_assets;
 pub mod check_package;
 pub mod fetch;
-pub mod gh_fetcher;
+pub mod github_fetcher;
 pub mod gitlab_fetcher;
 
 #[macro_use]
